backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
spacy==3.7.2
PyMuPDF==1.23.8
python-docx==0.8.11
Werkzeug==2.3.7